File Import
Imported file 'numInvest.py', a code written for PS 3 which numerically investigates the motion of a mass on a spring according to
the equation F = -kx. It will calculate the position and velocity at equally-spaced points for a specified time interval. It will also plot position and velocity as a function of time.